# [WIP] Proposal to enable webpack bundle optimization and html customization