--- a/dev-packages/application-manager/package.json
+++ b/dev-packages/application-manager/package.json
@@ -40,6 +40,7 @@
     "file-loader": "^1.1.11",
     "font-awesome-webpack": "0.0.5-beta.2",
     "fs-extra": "^4.0.2",
+    "html-webpack-plugin": "^3.2.0",
     "ignore-loader": "^0.1.2",
     "less": "^2.7.2",
     "source-map-loader": "^0.2.1",