### Pull Request Goal

The goal of this PR is to provide a way to:
- optionally optimize the packaging of bundles to make them caching-compliant, following some of the guidelines provided [here](https://webpack.js.org/guides/caching/).
- customize the generated `index.html`, by optionally providing a custom template to be used by the [Html Webpack Plugin](https://webpack.js.org/plugins/html-webpack-plugin/) and specifying additional template parameters. 

A typical use of these proposed features is to allow retrieving some of the generated bundle chunks from a CDN, and fallbacking to the Theia server when CDN files are not available.
 
### How it is done

#### Bundle optimization

Relevant [Webpack optimization options](https://webpack.js.org/configuration/optimization/) are added when the `--env-optimized` parameter is passed to the build.

#### HTML customization

- The [Html Webpack Plugin](https://webpack.js.org/plugins/html-webpack-plugin/) is now used in all cases, though producing the same output as before in the default case
- If a `custom-html-head.html` file is present in the Webpack build directory, its content is used inside the `head` tag of the Html template, instead of the default behavior that simply adds the `script` tags required to load the Javascript bundles (as was before).
In other words, when optimization is disabled, the default `head` tag of the generated Html would look like:
```
<head>
  <meta charset="UTF-8">
  <script src="bundle.js" type="text/javascript" charset="utf-8"></script>
</head>
``` 
But when the following `custom-html-head.html` file would be in the build folder:
```
  <script src="<%= htmlWebpackPlugin.options.someOptionsToProvideCDNURL %>" type="text/javascript" charset="utf-8"></script>
```
then the default `script` tag would be replaced by a line generated from this custom head template.
- The `--env--customtemplateparams` command-line build argument allows providing additional template parameters to be used by the `custom-html-head.html` template fragment, as a JSON string like this:
```
--env--customtemplateparams='{ "cdnPrefix": "some CDN URL" }'
```
Then, inside the template, this value can be used with the following code:
```
<%= htmlWebpackPlugin.options.customparams.cdnPrefix %>
```
 
#### Remarks

- This PR goes in the same direction as PR https://github.com/theia-ide/theia/pull/3398/files and seems to be somewhat a continuation of it.