@uniibu 2018-11-16
This line [#L69](https://github.com/davidfestal/theia/blob/544f3a000f7ff2e352df6ce36998398d25a4c86e/dev-packages/application-manager/src/application-package-manager.ts#L69) should be removed to avoid any conflict with `html-webpack-plugin` and besides, this pr renders that function useless.

Also instead of using custom templates for use with like cdn services, why not just let webpack do the generation as well. This would be less prone to dev error.
Something like:
```js
    output: {
        filename: optimized ? '[name].[contenthash].js' : 'bundle.js',
        path: optimized ? path.resolve(__dirname, 'cdn/assets') : outputPath,
	publicPath: optimized ? cdnPrefix : '' // cdnPrefix would be like https://cdn.example.com/assets/
    }
```
Then have `html-webpack-plugin` inject it on the head/body. Only caveat is that the dev will not be able to add custom external/remote script, but i think that is ideal for lesser prone dev error.