@davidfestal 2018-11-16
@uniibu 

> This line #L69 should be removed to avoid any conflict with html-webpack-plugin and besides, this pr renders that function useless.

Sure, I'll change this.

> Then have html-webpack-plugin inject it on the head/body.

well it doesn't seem flexible enough to allow:
- retrieving *only some of the files* from an alternate location
- retrieving the file from the Theia web server when it is not available on the CDN