@davidfestal 2018-11-16
> > There are 2 modes: production and development. In prod mode, Theia should be built optimized. What is the difference of `--env-optimized` to building in the production mode.
> 
> I think the reasoning for this is to make this PR optional, as stated by @geropl on the other [PR](https://github.com/theia-ide/theia/pull/3398#issuecomment-436529563), this might cause issues to the infrastructure where Theia is deployed to.

Exactly, the idea is that without this option the produced files will be the same as before, and nothing would be broken for existing users.
But I could check to disallow this option in development mode if you want @akosyakov