@davidfestal 2018-11-16
@akosyakov 

>If you do optimization, please bring numbers before/after for connections of different quality. One should be able to reproduce these numbers on his own to confirm that PR is a good optimization to go in the prod mode. If it is not good, but you still want to have it then use a custom webpack config for your app.

> Particularly to code splitting, we've tried it and it performed worse on slow connection than with one chunk.

Well there would be only 3 files there instead of one (I mean that's not *many* files).
I might also be that using `async` / `defer` options in the `script` tags might help leveraging concurrent downloads in the case of code splitting.

Anyway, this option is mainly a pre-requisite (or at least good-to-have) for long-term caching,  as explained in the mentioned Webpack documentation page, and has not much sense if caching is not planned. This is also why I made it optional. I could change the name of the option to `cacheable`, or something like that if it makes it clearer.