@akosyakov 2018-11-16
my point is there is already a way to add optional webpack configuration for its app, explained here: https://github.com/theia-ide/theia/issues/766#issuecomment-341720642

I don't see the point to add optional flags which are disabled by default next to it. If there is something which we can enable by default for prod/dev mode and everybody will for sure benefit from it, then it should be introduced and an optional flag should be added to disable it.