@uniibu 2018-11-16
> I don't see the point to add optional flags which are disabled by default next to it. If there is something which we can enable by default for prod/dev mode and everybody will for sure benefit from it, then it should be introduced and an optional flag should be added to disable it.

I do agree that if there is something which everyone/most could benefit,it should be introduced. However, based on a [comment](https://github.com/theia-ide/theia/pull/3398#issuecomment-436529563) on the previous/linked pr, Theia tends to lean more towards business/vendor usage/deployment. It is why this pr, even if it could benefit most, must be set optionally and disabled by default.

> my point is there is already a way to add optional webpack configuration for a Theia app, explained here: #766 (comment)

This is true for the most part. However customizations such as the goal of this PR can't be achieve solely by a custom webpack config. For one, this PR's goal for making a `caching-compliant` assets would be useless unless there is a way(correct me if im wrong) for a dev to set options to `express.static`(coz it always responds with `max-age:0`) without having to build a separate extension just for that purpose.