@uniibu 2018-11-16
Did some tests in regards to the time and speed.

**Using my own pc with cache disabled:**

Current(single bunde.js): load time of bundle.js **13.66s**
![pccurrent](https://i.gyazo.com/f98ae6da0f454e26a784d84c1af9a5ef.png)

Optimized(chunked): total load time of all chunked js **8.86s**
![pcchunked](https://i.gyazo.com/76c0f3901e66eb2005dbef48f3a8610e.png)

**Using GTMetrix(online)**

Current(single bundle.js): load time of bundle.js **8.68s**
![gtcurrent](https://i.gyazo.com/2661cbd5517ddbd8960f9518c45c55d7.png)

Optimized(chunked): total load time of all chunked js **4.15s**
![gtchunked](https://i.gyazo.com/2b5ddfd2cf6ba8a96d914f350b7e596d.png)

I did not add the test for `--split-frontend` or development mode as that takes too much time and latency to load.

The difference of this pr vs the current(single bundle.js) and `--split-frontend` is that the js files are already in-lined in the html file which means browsers can take advantage of parallel connections. Unlike the current or the `--split-frontend`, the js files are loaded programmatically via single js file,which doesn't take advantage of parallel connections and blocks rendering.