@davidfestal 2018-11-16
Thanks @uniibu for these tests, I'm not surprised of the results since I've been testing a POC on my side as well these last days.