@davidfestal 2018-11-19
> For one, this PR's goal for making a caching-compliant assets would be useless unless there is a way(correct me if im wrong) for a dev to set options to express.static(coz it always responds with max-age:0) without having to build a separate extension just for that purpose

@uniibu unless the caching-compliant files are then pushed to a CDN and retrieved from there through a customized `index.html` page, which is the solution I'm working on.