@davidfestal 2018-11-19
@akosyakov I'll try to use an additional webpack configuration on our side to avoid the changes in the Theia webpack configurator.

But are you OK with the html-related changes (use a custom head template if found in the build directory) ?